services/api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
elasticsearch==8.12.0
pydantic
requests
pytest
httpx
pandas
scikit-learn
joblib
psycopg2-binary
sqlalchemy
transformers
torch
